+++ projects/hipblaslt/library/src/amd_detail/hipblaslt-ext.cpp
@@ -27,6 +27,9 @@
 #include "hipblaslt/hipblaslt-ext.hpp"
 #include "exceptions.hpp"
 #include "hipblaslt_internal.hpp"
+#ifdef HIPBLASLT_ENABLE_JIT_GEMM
+#include "hipblaslt-jit-gemm-tag.hpp"
+#endif
 #include <Debug.hpp>
 #include <algorithm>
 #include <hip/hip_runtime.h>
@@ -1484,6 +1487,11 @@ namespace hipblaslt_ext
 
     int getIndexFromAlgo(hipblasLtMatmulAlgo_t& algo)
     {
+#ifdef HIPBLASLT_ENABLE_JIT_GEMM
+        if(experimental::detail::isJitAlgo(
+               *reinterpret_cast<const rocblaslt_matmul_algo*>(&algo)))
+            return -1;
+#endif
         int* algo_ptr = (int*)algo.data;
 
         return *algo_ptr;
@@ -1586,4 +1594,31 @@ namespace hipblaslt_ext
         }
         return HIPBLAS_STATUS_SUCCESS;
     }
+#ifndef HIPBLASLT_ENABLE_JIT_GEMM
+    namespace experimental
+    {
+        hipblasStatus_t getJitGemmAlgo(hipblasLtHandle_t,
+                                       hipblasLtMatmulDesc_t,
+                                       const void*,
+                                       const void*,
+                                       hipblasLtMatrixLayout_t,
+                                       const void*,
+                                       hipblasLtMatrixLayout_t,
+                                       const void*,
+                                       const void*,
+                                       hipblasLtMatrixLayout_t,
+                                       void*,
+                                       hipblasLtMatrixLayout_t,
+                                       const GenerateOptions&,
+                                       size_t,
+                                       hipblasLtMatmulHeuristicResult_t& result,
+                                       JitGemmInfo&                      info)
+        {
+            result     = {};
+            info       = {};
+            info.error = "JIT GEMM requires HIPBLASLT_ENABLE_JIT_GEMM=ON";
+            return HIPBLAS_STATUS_NOT_SUPPORTED;
+        }
+    }
+#endif
 } // End of namespace hipblasltext